Refactor SieveListCollector to a single class
